SciViews-K added (and SciViews-R eliminated)
SciViews-K is now on Github. Added to the list of add ons, and the link to the old version eliminated.
I had also to eliminate the reference to SciViews-R (svKomodo R package) that someone else added to this list. Indeed, svKomodo is an R package I wrote myself to interact with Komodo from R. It is **not** a Komodo add ons. Thus, it should not appear on this list!